requirements.txt
simplejson
faster-whisper
python-dotenv
transformers
requests
spotipy